A mimic of collocate

Create a sparse feature co-occurrence matrix. This function is a shorthand of pipian::gbs_dfm() and then quanteda::fcm().

gbs_collocate(df, pull = "token", n = 1L, sep = "-", what = "fastestword", ...)

Arguments

df
Prettified data.frame of tokens.

pull
Column to be packed into text or ngrams body. Default value is `token`.

n
Integer passed to quanteda::char_ngrams.

sep
Character scalar internally used as the concatenator of ngrams inside quanteda::char_ngrams.

what
Character scalar. Which tokenizer to use in quanteda::tokens().

...
Other arguments are passed to quanteda::tokens().

Value

quanteda fcm object

See also