Pack prettified data.frame of tokens

Pack a prettified data.frame of tokens into a new data.frame of corpus, which is compatible with the Text Interchange Formats.

pack(df, n = 1L, skip = 0L, pull = "token", sep = "_", .collapse = " ")

Arguments

df
A prettified data.frame of tokens.

n
Integer passed to quanteda::char_ngrams().

skip
Integer passed to quanteda::char_ngrams().

pull
Column to be packed into text or ngrams body. Default value is `token`.

sep
Character scalar internally used as the concatenator of ngrams inside quanteda::char_ngrams.

.collapse
This argument is passed to stringi::stri_c().

Value

A data.frame.

Text Interchange Formats (TIF)

The Text Interchange Formats (TIF) is a set of standards that allows R text analysis packages to target defined inputs and outputs for corpora, tokens, and document-term matrices.

Valid data.frame of tokens

The prettified data.frame of tokens here is a data.frame object compatible with the TIF.

A TIF valid data.frame of tokens are expected to have one unique key column (named `doc_id`) of each text and several feature columns of each tokens. The feature columns must contain at least `token` itself.

See also

Examples

xml <- ppn_parse_xml(system.file("sample.xml", package = "pipian"))
head(pack(xml))

#>   doc_id
#> 1      1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      text
#> 1 <U+3042><U+308B> <U+65E5> <U+306E> <U+4E8B> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+5FA1> <U+91C8><U+8FE6> <U+69D8> <U+306F> <U+6975><U+697D> <U+306E> <U+84EE><U+6C60> <U+306E> <U+3075><U+3061> <U+3092> <U+3001> <U+72EC><U+308A> <U+3067> <U+3076><U+3089><U+3076><U+3089> <U+5FA1> <U+6B69><U+304D> <U+306B> <U+306A><U+3063> <U+3066> <U+3044><U+3089><U+3063><U+3057><U+3083><U+3044> <U+307E><U+3057> <U+305F> <U+3002> <U+6C60> <U+306E> <U+4E2D> <U+306B> <U+54B2><U+3044> <U+3066> <U+3044><U+308B> <U+84EE> <U+306E> <U+82B1> <U+306F> <U+3001> <U+307F><U+3093><U+306A> <U+7389> <U+306E> <U+3088><U+3046> <U+306B> <U+307E><U+3063><U+767D> <U+3067> <U+3001> <U+305D><U+306E> <U+307E><U+3093><U+4E2D> <U+306B> <U+3042><U+308B> <U+91D1><U+8272> <U+306E> <U+854A> <U+304B><U+3089> <U+306F> <U+3001> <U+4F55><U+3068><U+3082> <U+4E91><U+3048> <U+306A><U+3044> <U+597D><U+3044> <U+5302> <U+304C> <U+3001> <U+7D76><U+9593> <U+306A><U+304F> <U+3042><U+305F><U+308A> <U+3078> <U+6EA2><U+308C> <U+3066> <U+5C45><U+308A> <U+307E><U+3059> <U+3002> <U+6975><U+697D> <U+306F> <U+4E01><U+5EA6> <U+671D> <U+306A> <U+306E> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3057><U+3087> <U+3046> <U+3002> <U+3084><U+304C><U+3066> <U+5FA1> <U+91C8><U+8FE6> <U+69D8> <U+306F> <U+305D><U+306E> <U+6C60> <U+306E> <U+3075><U+3061> <U+306B> <U+5FA1> <U+4F47><U+307F> <U+306B> <U+306A><U+3063> <U+3066> <U+3001> <U+6C34> <U+306E> <U+9762> <U+3092> <U+853D><U+3063> <U+3066> <U+3044><U+308B> <U+84EE> <U+306E> <U+8449> <U+306E> <U+9593> <U+304B><U+3089> <U+3001> <U+3075><U+3068> <U+4E0B> <U+306E> <U+5BB9><U+5B50> <U+3092> <U+5FA1><U+89A7> <U+306B> <U+306A><U+308A> <U+307E><U+3057> <U+305F> <U+3002> <U+3053><U+306E> <U+6975><U+697D> <U+306E> <U+84EE><U+6C60> <U+306E> <U+4E0B> <U+306F> <U+3001> <U+4E01><U+5EA6> <U+5730><U+7344> <U+306E> <U+5E95> <U+306B> <U+5F53><U+3063> <U+3066> <U+5C45><U+308A> <U+307E><U+3059> <U+304B><U+3089> <U+3001> <U+6C34><U+6676> <U+306E> <U+3088><U+3046> <U+306A> <U+6C34> <U+3092> <U+900F><U+304D> <U+5FB9><U+3057> <U+3066> <U+3001> <U+4E09> <U+9014> <U+306E> <U+6CB3> <U+3084> <U+91DD> <U+306E> <U+5C71> <U+306E> <U+666F><U+8272> <U+304C> <U+3001> <U+4E01><U+5EA6> <U+8997><U+304D> <U+773C><U+93E1> <U+3092> <U+898B><U+308B> <U+3088><U+3046> <U+306B> <U+3001> <U+306F><U+3063><U+304D><U+308A> <U+3068> <U+898B><U+3048><U+308B> <U+306E> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+3059><U+308B><U+3068> <U+305D><U+306E> <U+5730><U+7344> <U+306E> <U+5E95> <U+306B> <U+3001> <U+9640><U+591A> <U+3068> <U+4E91><U+3046> <U+7537> <U+304C> <U+4E00> <U+4EBA> <U+3001> <U+307B><U+304B> <U+306E> <U+7F6A><U+4EBA> <U+3068> <U+4E00> <U+3057><U+3087> <U+306B> <U+8822><U+3044> <U+3066> <U+3044><U+308B> <U+59FF> <U+304C> <U+3001> <U+5FA1> <U+773C> <U+306B> <U+6B62><U+307E><U+308A> <U+307E><U+3057> <U+305F> <U+3002> <U+3053><U+306E> <U+9640><U+591A> <U+3068> <U+4E91><U+3046> <U+7537> <U+306F> <U+3001> <U+4EBA> <U+3092> <U+6BBA><U+3057> <U+305F><U+308A> <U+5BB6> <U+306B> <U+706B> <U+3092> <U+3064><U+3051> <U+305F><U+308A> <U+3001> <U+3044><U+308D><U+3044><U+308D> <U+60AA><U+4E8B> <U+3092> <U+50CD><U+3044> <U+305F> <U+5927> <U+6CE5><U+574A> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+304C> <U+3001> <U+305D><U+308C><U+3067><U+3082> <U+305F><U+3063><U+305F> <U+4E00><U+3064> <U+3001> <U+5584><U+3044> <U+4E8B> <U+3092> <U+81F4><U+3057> <U+305F> <U+899A><U+3048> <U+304C> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+3068> <U+7533><U+3057> <U+307E><U+3059> <U+306E> <U+306F> <U+3001> <U+3042><U+308B> <U+6642> <U+3053><U+306E> <U+7537> <U+304C> <U+6DF1><U+3044> <U+6797> <U+306E> <U+4E2D> <U+3092> <U+901A><U+308A> <U+307E><U+3059> <U+3068> <U+3001> <U+5C0F><U+3055><U+306A> <U+8718><U+86DB> <U+304C> <U+4E00> <U+5339> <U+3001> <U+8DEF> <U+3070> <U+305F> <U+3092> <U+9019><U+3063> <U+3066> <U+884C><U+304F> <U+306E> <U+304C> <U+898B><U+3048> <U+307E><U+3057> <U+305F> <U+3002> <U+305D><U+3053><U+3067> <U+9640><U+591A> <U+306F> <U+65E9><U+901F> <U+8DB3> <U+3092> <U+6319><U+3052> <U+3066> <U+3001> <U+8E0F><U+307F> <U+6BBA><U+305D> <U+3046> <U+3068> <U+81F4><U+3057> <U+307E><U+3057> <U+305F> <U+304C> <U+3001> <U+300C> <U+3044><U+3084> <U+3001> <U+3044><U+3084> <U+3001> <U+3053><U+308C> <U+3082> <U+5C0F><U+3055><U+3044> <U+306A><U+304C><U+3089> <U+3001> <U+547D> <U+306E> <U+3042><U+308B> <U+3082><U+306E> <U+306B> <U+9055><U+3044> <U+306A><U+3044> <U+3002> <U+305D><U+306E> <U+547D> <U+3092> <U+7121> <U+6697><U+306B> <U+3068><U+308B> <U+3068> <U+4E91><U+3046> <U+4E8B> <U+306F> <U+3001> <U+3044><U+304F><U+3089> <U+4F55> <U+3067><U+3082> <U+53EF><U+54C0><U+305D><U+3046> <U+3060> <U+3002> <U+300D> <U+3068> <U+3001> <U+3053><U+3046> <U+6025> <U+306B> <U+601D><U+3044><U+8FD4><U+3057> <U+3066> <U+3001> <U+3068><U+3046><U+3068><U+3046> <U+305D><U+306E> <U+8718><U+86DB> <U+3092> <U+6BBA><U+3055> <U+305A> <U+306B> <U+52A9><U+3051> <U+3066> <U+3084><U+3063> <U+305F> <U+304B><U+3089> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+5FA1> <U+91C8><U+8FE6> <U+69D8> <U+306F> <U+5730><U+7344> <U+306E> <U+5BB9><U+5B50> <U+3092> <U+5FA1><U+89A7> <U+306B> <U+306A><U+308A> <U+306A><U+304C><U+3089> <U+3001> <U+3053><U+306E> <U+9640><U+591A> <U+306B> <U+306F> <U+8718><U+86DB> <U+3092> <U+52A9><U+3051> <U+305F> <U+4E8B> <U+304C> <U+3042><U+308B> <U+306E> <U+3092> <U+5FA1> <U+601D><U+3044><U+51FA><U+3057> <U+306B> <U+306A><U+308A> <U+307E><U+3057> <U+305F> <U+3002> <U+305D><U+3046><U+3057><U+3066> <U+305D><U+308C><U+3060><U+3051> <U+306E> <U+5584><U+3044> <U+4E8B> <U+3092> <U+3057> <U+305F> <U+5831> <U+306B> <U+306F> <U+3001> <U+51FA><U+6765><U+308B> <U+306A><U+3089> <U+3001> <U+3053><U+306E> <U+7537> <U+3092> <U+5730><U+7344> <U+304B><U+3089> <U+6551><U+3044><U+51FA><U+3057> <U+3066> <U+3084><U+308D> <U+3046> <U+3068> <U+5FA1> <U+8003><U+3048> <U+306B> <U+306A><U+308A> <U+307E><U+3057> <U+305F> <U+3002> <U+5E78><U+3044> <U+3001> <U+5074> <U+3092> <U+898B> <U+307E><U+3059> <U+3068> <U+3001> <U+7FE1><U+7FE0> <U+306E> <U+3088><U+3046> <U+306A> <U+8272> <U+3092> <U+3057> <U+305F> <U+84EE> <U+306E> <U+8449> <U+306E> <U+4E0A> <U+306B> <U+3001> <U+6975><U+697D> <U+306E> <U+8718><U+86DB> <U+304C> <U+4E00> <U+5339> <U+3001> <U+7F8E><U+3057><U+3044> <U+9280><U+8272> <U+306E> <U+7CF8> <U+3092> <U+304B><U+3051> <U+3066> <U+5C45><U+308A> <U+307E><U+3059> <U+3002> <U+5FA1> <U+91C8><U+8FE6> <U+69D8> <U+306F> <U+305D><U+306E> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+3092> <U+305D><U+3063><U+3068> <U+5FA1> <U+624B> <U+306B> <U+5FA1> <U+53D6><U+308A> <U+306B> <U+306A><U+3063> <U+3066> <U+3001> <U+7389> <U+306E> <U+3088><U+3046> <U+306A> <U+767D><U+84EE> <U+306E> <U+9593> <U+304B><U+3089> <U+3001> <U+9065><U+304B> <U+4E0B> <U+306B> <U+3042><U+308B> <U+5730><U+7344> <U+306E> <U+5E95> <U+3078> <U+3001> <U+307E><U+3063><U+3059><U+3050> <U+306B> <U+305D><U+308C> <U+3092> <U+5FA1> <U+4E0B><U+3057> <U+306A><U+3055><U+3044> <U+307E><U+3057> <U+305F> <U+3002> <U+3053><U+3061><U+3089> <U+306F> <U+5730><U+7344> <U+306E> <U+5E95> <U+306E> <U+8840><U+306E><U+6C60> <U+3067> <U+3001> <U+307B><U+304B> <U+306E> <U+7F6A><U+4EBA> <U+3068> <U+4E00> <U+3057><U+3087> <U+306B> <U+3001> <U+6D6E><U+3044> <U+305F><U+308A> <U+6C88><U+3093> <U+3060><U+308A> <U+3057> <U+3066> <U+3044> <U+305F> <U+9640><U+591A> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+4F55><U+3057><U+308D> <U+3069><U+3061><U+3089> <U+3092> <U+898B> <U+3066> <U+3082> <U+3001> <U+307E><U+3063><U+6697> <U+3067> <U+3001> <U+305F><U+307E><U+306B> <U+305D><U+306E> <U+304F><U+3089> <U+6697><U+304B><U+3089> <U+307C><U+3093><U+3084><U+308A> <U+6D6E><U+304D> <U+4E0A><U+3063> <U+3066> <U+3044><U+308B> <U+3082><U+306E> <U+304C> <U+3042><U+308B> <U+3068> <U+601D><U+3044> <U+307E><U+3059> <U+3068> <U+3001> <U+305D><U+308C> <U+306F> <U+6050><U+3057> <U+3044> <U+91DD> <U+306E> <U+5C71> <U+306E> <U+91DD> <U+304C> <U+5149><U+308B> <U+306E> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+304B><U+3089> <U+3001> <U+305D><U+306E> <U+5FC3><U+7D30> <U+3055> <U+3068> <U+4E91><U+3063> <U+305F><U+3089> <U+3054><U+3056><U+3044> <U+307E><U+305B> <U+3093> <U+3002> <U+305D><U+306E> <U+4E0A> <U+3042><U+305F><U+308A> <U+306F> <U+5893> <U+306E> <U+4E2D> <U+306E> <U+3088><U+3046> <U+306B> <U+3057><U+3093> <U+3068> <U+9759><U+307E><U+308A><U+8FD4><U+3063> <U+3066> <U+3001> <U+305F><U+307E> <U+306B> <U+805E><U+3048><U+308B> <U+3082><U+306E> <U+3068> <U+4E91><U+3063> <U+3066> <U+306F> <U+3001> <U+305F><U+3060> <U+7F6A><U+4EBA> <U+304C> <U+3064><U+304F> <U+5FAE> <U+306A> <U+5606><U+606F> <U+3070><U+304B><U+308A> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+3053><U+308C> <U+306F> <U+3053><U+3053> <U+3078> <U+843D><U+3061> <U+3066> <U+6765><U+308B> <U+307B><U+3069> <U+306E> <U+4EBA><U+9593> <U+306F> <U+3001> <U+3082><U+3046> <U+3055><U+307E><U+3056><U+307E> <U+306A> <U+5730><U+7344> <U+306E> <U+8CAC><U+82E6> <U+306B> <U+75B2><U+308C><U+306F><U+3066> <U+3066> <U+3001> <U+6CE3><U+58F0> <U+3092> <U+51FA><U+3059> <U+529B> <U+3055><U+3048> <U+306A><U+304F><U+306A><U+3063> <U+3066> <U+3044><U+308B> <U+306E> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3057><U+3087> <U+3046> <U+3002> <U+3067><U+3059><U+304B><U+3089> <U+3055><U+3059><U+304C> <U+5927> <U+6CE5><U+574A> <U+306E> <U+9640><U+591A> <U+3082> <U+3001> <U+3084><U+306F><U+308A> <U+8840><U+306E><U+6C60> <U+306E> <U+8840> <U+306B> <U+54BD><U+3073> <U+306A><U+304C><U+3089> <U+3001> <U+307E><U+308B><U+3067> <U+6B7B><U+306B> <U+304B><U+304B><U+3063> <U+305F> <U+86D9> <U+306E> <U+3088><U+3046> <U+306B> <U+3001> <U+305F><U+3060> <U+3082><U+304C><U+3044> <U+3066> <U+3070><U+304B><U+308A> <U+5C45><U+308A> <U+307E><U+3057> <U+305F> <U+3002> <U+3068><U+3053><U+308D><U+304C> <U+3042><U+308B> <U+6642> <U+306E> <U+4E8B> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+4F55><U+6C17><U+306A><U+304F> <U+9640><U+591A> <U+304C> <U+982D> <U+3092> <U+6319><U+3052> <U+3066> <U+3001> <U+8840><U+306E><U+6C60> <U+306E> <U+7A7A> <U+3092> <U+773A><U+3081> <U+307E><U+3059> <U+3068> <U+3001> <U+305D><U+306E> <U+3072><U+3063><U+305D><U+308A> <U+3068> <U+3057> <U+305F> <U+6697> <U+306E> <U+4E2D> <U+3092> <U+3001> <U+9060><U+3044> <U+9060><U+3044> <U+5929><U+4E0A> <U+304B><U+3089> <U+3001> <U+9280><U+8272> <U+306E> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+304C> <U+3001> <U+307E><U+308B><U+3067> <U+4EBA><U+76EE> <U+306B> <U+304B><U+304B><U+308B> <U+306E> <U+3092> <U+6050><U+308C><U+308B> <U+3088><U+3046> <U+306B> <U+3001> <U+4E00><U+3059><U+3058> <U+7D30><U+304F> <U+5149><U+308A> <U+306A><U+304C><U+3089> <U+3001> <U+3059><U+308B><U+3059><U+308B> <U+3068> <U+81EA><U+5206> <U+306E> <U+4E0A> <U+3078> <U+5782><U+308C> <U+3066> <U+53C2><U+308B> <U+306E> <U+3067> <U+306F> <U+3054><U+3056><U+3044> <U+307E><U+305B> <U+3093> <U+304B> <U+3002> <U+9640><U+591A> <U+306F> <U+3053><U+308C> <U+3092> <U+898B><U+308B> <U+3068> <U+3001> <U+601D><U+308F><U+305A> <U+624B> <U+3092> <U+62CD> <U+3063><U+3066> <U+559C><U+3073> <U+307E><U+3057> <U+305F> <U+3002> <U+3053><U+306E> <U+7CF8> <U+306B> <U+7E0B><U+308A> <U+3064><U+3044> <U+3066> <U+3001> <U+3069><U+3053> <U+307E><U+3067> <U+3082> <U+306E><U+307C><U+3063> <U+3066> <U+884C><U+3051> <U+3070> <U+3001> <U+304D><U+3063><U+3068> <U+5730><U+7344> <U+304B><U+3089> <U+306C><U+3051> <U+51FA><U+305B><U+308B> <U+306E> <U+306B> <U+76F8><U+9055> <U+3054><U+3056><U+3044> <U+307E><U+305B> <U+3093> <U+3002> <U+3044><U+3084> <U+3001> <U+3046><U+307E><U+304F> <U+884C><U+304F> <U+3068> <U+3001> <U+6975><U+697D> <U+3078> <U+306F><U+3044><U+308B> <U+4E8B> <U+3055><U+3048> <U+3082> <U+51FA><U+6765> <U+307E><U+3057><U+3087> <U+3046> <U+3002> <U+305D><U+3046> <U+3059><U+308C> <U+3070> <U+3001> <U+3082><U+3046> <U+91DD> <U+306E> <U+5C71> <U+3078> <U+8FFD><U+3044><U+4E0A><U+3052> <U+3089><U+308C><U+308B> <U+4E8B> <U+3082> <U+306A><U+304F><U+306A><U+308C> <U+3070> <U+3001> <U+8840><U+306E><U+6C60> <U+306B> <U+6C88><U+3081> <U+3089><U+308C><U+308B> <U+4E8B> <U+3082> <U+3042><U+308B> <U+7B48> <U+306F> <U+3054><U+3056><U+3044> <U+307E><U+305B> <U+3093> <U+3002> <U+3053><U+3046> <U+601D><U+3044> <U+307E><U+3057> <U+305F> <U+304B><U+3089> <U+9640><U+591A> <U+306F> <U+3001> <U+65E9><U+901F> <U+305D><U+306E> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+3092> <U+4E21><U+624B> <U+3067> <U+3057><U+3063><U+304B><U+308A> <U+3068> <U+3064><U+304B><U+307F> <U+306A><U+304C><U+3089> <U+3001> <U+4E00><U+751F><U+61F8><U+547D> <U+306B> <U+4E0A> <U+3078> <U+4E0A> <U+3078> <U+3068> <U+305F><U+3050><U+308A> <U+306E><U+307C><U+308A> <U+59CB><U+3081> <U+307E><U+3057> <U+305F> <U+3002> <U+5143> <U+3088><U+308A> <U+5927> <U+6CE5><U+574A> <U+306E> <U+4E8B> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+304B><U+3089> <U+3001> <U+3053><U+3046> <U+4E91><U+3046> <U+4E8B> <U+306B> <U+306F> <U+6614> <U+304B><U+3089> <U+3001> <U+6163><U+308C> <U+5207><U+3063> <U+3066> <U+3044><U+308B> <U+306E> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+3057><U+304B><U+3057> <U+5730><U+7344> <U+3068> <U+6975><U+697D> <U+3068> <U+306E> <U+9593> <U+306F> <U+3001> <U+4F55> <U+4E07><U+91CC> <U+3068> <U+306A><U+304F> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+304B><U+3089> <U+3001> <U+3044><U+304F><U+3089> <U+7126><U+3063> <U+3066> <U+898B> <U+305F> <U+6240> <U+3067> <U+3001> <U+5BB9><U+6613> <U+306B> <U+4E0A> <U+3078> <U+306F> <U+51FA> <U+3089><U+308C> <U+307E><U+305B> <U+3093> <U+3002> <U+3084><U+3084> <U+3057><U+3070><U+3089><U+304F> <U+306E><U+307C><U+308B> <U+4E2D> <U+306B> <U+3001> <U+3068><U+3046><U+3068><U+3046> <U+9640><U+591A> <U+3082> <U+304F><U+305F><U+3073><U+308C> <U+3066> <U+3001> <U+3082><U+3046> <U+4E00> <U+305F><U+3050><U+308A> <U+3082> <U+4E0A> <U+306E> <U+65B9> <U+3078> <U+306F> <U+306E><U+307C><U+308C> <U+306A><U+304F> <U+306A><U+3063> <U+3066> <U+3057><U+307E><U+3044> <U+307E><U+3057> <U+305F> <U+3002> <U+305D><U+3053><U+3067> <U+4ED5><U+65B9> <U+304C> <U+3054><U+3056><U+3044> <U+307E><U+305B> <U+3093> <U+304B><U+3089> <U+3001> <U+307E><U+305A> <U+4E00><U+4F11><U+307F> <U+4F11><U+3080> <U+3064><U+3082><U+308A> <U+3067> <U+3001> <U+7CF8> <U+306E> <U+4E2D><U+9014> <U+306B> <U+3076><U+3089> <U+4E0B><U+308A> <U+306A><U+304C><U+3089> <U+3001> <U+9065><U+304B> <U+306B> <U+76EE><U+306E><U+4E0B> <U+3092> <U+898B><U+4E0B><U+3057> <U+307E><U+3057> <U+305F> <U+3002> <U+3059><U+308B><U+3068> <U+3001> <U+4E00><U+751F><U+61F8><U+547D> <U+306B> <U+306E><U+307C><U+3063> <U+305F> <U+7532><U+6590> <U+304C> <U+3042><U+3063> <U+3066> <U+3001> <U+3055><U+3063><U+304D> <U+307E><U+3067> <U+81EA><U+5206> <U+304C> <U+3044> <U+305F> <U+8840><U+306E><U+6C60> <U+306F> <U+3001> <U+4ECA> <U+3067> <U+306F> <U+3082><U+3046> <U+6697> <U+306E> <U+5E95> <U+306B> <U+3044><U+3064><U+306E><U+9593><U+306B><U+304B> <U+304B><U+304F><U+308C> <U+3066> <U+5C45><U+308A> <U+307E><U+3059> <U+3002> <U+305D><U+308C><U+304B><U+3089> <U+3042><U+306E> <U+307C><U+3093><U+3084><U+308A> <U+5149><U+3063> <U+3066> <U+3044><U+308B> <U+6050><U+3057> <U+3044> <U+91DD> <U+306E> <U+5C71> <U+3082> <U+3001> <U+8DB3> <U+306E> <U+4E0B> <U+306B> <U+306A><U+3063> <U+3066> <U+3057><U+307E><U+3044> <U+307E><U+3057> <U+305F> <U+3002> <U+3053><U+306E> <U+5206> <U+3067> <U+306E><U+307C><U+3063> <U+3066> <U+884C><U+3051> <U+3070> <U+3001> <U+5730><U+7344> <U+304B><U+3089> <U+306C><U+3051> <U+51FA><U+3059> <U+306E> <U+3082> <U+3001> <U+5B58><U+5916> <U+308F><U+3051> <U+304C> <U+306A><U+3044> <U+304B><U+3082> <U+77E5><U+308C> <U+307E><U+305B> <U+3093> <U+3002> <U+9640><U+591A> <U+306F> <U+4E21><U+624B> <U+3092> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+306B> <U+304B><U+3089><U+307F> <U+306A><U+304C><U+3089> <U+3001> <U+3053><U+3053> <U+3078> <U+6765> <U+3066> <U+304B><U+3089> <U+4F55> <U+5E74> <U+306B> <U+3082> <U+51FA><U+3057> <U+305F> <U+4E8B> <U+306E> <U+306A><U+3044> <U+58F0> <U+3067> <U+3001> <U+300C> <U+3057><U+3081> <U+305F> <U+3002> <U+3057><U+3081> <U+305F> <U+3002> <U+300D> <U+3068> <U+7B11><U+3044> <U+307E><U+3057> <U+305F> <U+3002> <U+3068><U+3053><U+308D><U+304C> <U+3075><U+3068> <U+6C17> <U+304C> <U+3064><U+304D> <U+307E><U+3059> <U+3068> <U+3001> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+306E> <U+4E0B> <U+306E> <U+65B9> <U+306B> <U+306F> <U+3001> <U+6570> <U+9650> <U+3082> <U+306A><U+3044> <U+7F6A><U+4EBA> <U+305F><U+3061> <U+304C> <U+3001> <U+81EA><U+5206> <U+306E> <U+306E><U+307C><U+3063> <U+305F> <U+5F8C> <U+3092> <U+3064><U+3051> <U+3066> <U+3001> <U+307E><U+308B><U+3067> <U+87FB> <U+306E> <U+884C><U+5217> <U+306E> <U+3088><U+3046> <U+306B> <U+3001> <U+3084><U+306F><U+308A> <U+4E0A> <U+3078> <U+4E0A> <U+3078> <U+4E00><U+5FC3> <U+306B> <U+3088><U+3058><U+306E><U+307C><U+3063> <U+3066> <U+6765><U+308B> <U+3067> <U+306F> <U+3054><U+3056><U+3044> <U+307E><U+305B> <U+3093> <U+304B> <U+3002> <U+9640><U+591A> <U+306F> <U+3053><U+308C> <U+3092> <U+898B><U+308B> <U+3068> <U+3001> <U+9A5A><U+3044> <U+305F> <U+306E> <U+3068> <U+6050> <U+3057> <U+3044> <U+306E> <U+3068> <U+3067> <U+3001> <U+3057><U+3070><U+3089><U+304F> <U+306F> <U+305F><U+3060> <U+3001> <U+83AB><U+8FE6> <U+306E> <U+3088><U+3046> <U+306B> <U+5927><U+304D><U+306A> <U+53E3> <U+3092> <U+958B><U+3044> <U+305F> <U+307E><U+307E> <U+3001> <U+773C> <U+3070><U+304B><U+308A> <U+52D5><U+304B><U+3057> <U+3066> <U+5C45><U+308A> <U+307E><U+3057> <U+305F> <U+3002> <U+81EA><U+5206> <U+4E00> <U+4EBA> <U+3067> <U+3055><U+3048> <U+65AD><U+308C> <U+305D><U+3046> <U+306A> <U+3001> <U+3053><U+306E> <U+7D30><U+3044> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+304C> <U+3001> <U+3069><U+3046><U+3057><U+3066> <U+3042><U+308C> <U+3060><U+3051> <U+306E> <U+4EBA><U+6570> <U+306E> <U+91CD><U+307F> <U+306B> <U+582A><U+3048><U+308B> <U+4E8B> <U+304C> <U+51FA><U+6765> <U+307E><U+3057><U+3087> <U+3046> <U+3002> <U+3082><U+3057> <U+4E07><U+4E00> <U+9014><U+4E2D> <U+3067> <U+65AD><U+308C> <U+305F> <U+3068> <U+81F4><U+3057> <U+307E><U+3057> <U+305F><U+3089> <U+3001> <U+6298><U+89D2> <U+3053><U+3053> <U+3078> <U+307E><U+3067> <U+306E><U+307C><U+3063> <U+3066> <U+6765> <U+305F> <U+3053><U+306E> <U+809D><U+814E> <U+306A> <U+81EA><U+5206> <U+307E><U+3067> <U+3082> <U+3001> <U+5143> <U+306E> <U+5730><U+7344> <U+3078> <U+9006><U+843D><U+3057> <U+306B> <U+843D><U+3061> <U+3066> <U+3057><U+307E><U+308F> <U+306A><U+3051><U+308C> <U+3070> <U+306A><U+308A> <U+307E><U+305B> <U+3093> <U+3002> <U+305D><U+3093><U+306A> <U+4E8B> <U+304C> <U+3042><U+3063> <U+305F><U+3089> <U+3001> <U+5927><U+5909> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+304C> <U+3001> <U+305D><U+3046> <U+4E91><U+3046> <U+4E2D> <U+306B> <U+3082> <U+3001> <U+7F6A><U+4EBA> <U+305F><U+3061> <U+306F> <U+4F55> <U+767E> <U+3068> <U+306A><U+304F> <U+4F55> <U+5343> <U+3068> <U+306A><U+304F> <U+3001> <U+307E><U+3063><U+6697> <U+306A> <U+8840><U+306E><U+6C60> <U+306E> <U+5E95> <U+304B><U+3089> <U+3001> <U+3046><U+3088><U+3046><U+3088> <U+3068> <U+9019><U+3044> <U+4E0A><U+3063> <U+3066> <U+3001> <U+7D30><U+304F> <U+5149><U+3063> <U+3066> <U+3044><U+308B> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+3092> <U+3001> <U+4E00> <U+5217> <U+306B> <U+306A><U+308A> <U+306A><U+304C><U+3089> <U+3001> <U+305B><U+3063><U+305B><U+3068> <U+306E><U+307C><U+3063> <U+3066> <U+53C2><U+308A> <U+307E><U+3059> <U+3002> <U+4ECA> <U+306E> <U+4E2D> <U+306B> <U+3069><U+3046><U+304B> <U+3057> <U+306A><U+3051><U+308C> <U+3070> <U+3001> <U+7CF8> <U+306F> <U+307E><U+3093><U+4E2D> <U+304B><U+3089> <U+4E8C><U+3064> <U+306B> <U+65AD><U+308C> <U+3066> <U+3001> <U+843D><U+3061> <U+3066> <U+3057><U+307E><U+3046> <U+306E> <U+306B> <U+9055><U+3044> <U+3042><U+308A> <U+307E><U+305B> <U+3093> <U+3002> <U+305D><U+3053><U+3067> <U+9640><U+591A> <U+306F> <U+5927><U+304D><U+306A> <U+58F0> <U+3092> <U+51FA><U+3057> <U+3066> <U+3001> <U+300C> <U+3053><U+3089> <U+3001> <U+7F6A><U+4EBA> <U+3069><U+3082> <U+3002> <U+3053><U+306E> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+306F> <U+5DF1> <U+306E> <U+3082><U+306E> <U+3060> <U+305E> <U+3002> <U+304A><U+524D> <U+305F><U+3061> <U+306F> <U+4E00><U+4F53> <U+8AB0> <U+306B> <U+5C0B> <U+3044> <U+3066> <U+3001> <U+306E><U+307C><U+3063> <U+3066> <U+6765> <U+305F> <U+3002> <U+4E0B><U+308A><U+308D> <U+3002> <U+4E0B><U+308A><U+308D> <U+3002> <U+300D> <U+3068> <U+559A><U+304D> <U+307E><U+3057> <U+305F> <U+3002> <U+305D><U+306E> <U+9014><U+7AEF> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+4ECA> <U+307E><U+3067> <U+4F55> <U+3068> <U+3082> <U+306A><U+304B><U+3063> <U+305F> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+304C> <U+3001> <U+6025> <U+306B> <U+9640><U+591A> <U+306E> <U+3076><U+3089> <U+4E0B><U+3063> <U+3066> <U+3044><U+308B> <U+6240> <U+304B><U+3089> <U+3001> <U+3077><U+3064><U+308A><U+3068> <U+97F3> <U+3092> <U+7ACB><U+3066> <U+3066> <U+65AD><U+308C> <U+307E><U+3057> <U+305F> <U+3002> <U+3067><U+3059><U+304B><U+3089> <U+9640><U+591A> <U+3082> <U+305F><U+307E><U+308A> <U+307E><U+305B> <U+3093> <U+3002> <U+3042><U+3063> <U+3068> <U+4E91><U+3046> <U+9593><U+3082><U+306A><U+304F> <U+98A8> <U+3092> <U+5207><U+3063> <U+3066> <U+3001> <U+72EC><U+697D> <U+306E> <U+3088><U+3046> <U+306B> <U+304F><U+308B><U+304F><U+308B> <U+307E><U+308F><U+308A> <U+306A><U+304C><U+3089> <U+3001> <U+898B><U+308B> <U+898B><U+308B> <U+4E2D> <U+306B> <U+6697> <U+306E> <U+5E95> <U+3078> <U+3001> <U+307E><U+3063> <U+3055> <U+304B> <U+3055><U+307E> <U+306B> <U+843D><U+3061> <U+3066> <U+3057><U+307E><U+3044> <U+307E><U+3057> <U+305F> <U+3002> <U+5F8C> <U+306B> <U+306F> <U+305F><U+3060> <U+6975><U+697D> <U+306E> <U+8718><U+86DB> <U+306E> <U+7CF8> <U+304C> <U+3001> <U+304D><U+3089><U+304D><U+3089> <U+3068> <U+7D30><U+304F> <U+5149><U+308A> <U+306A><U+304C><U+3089> <U+3001> <U+6708> <U+3082> <U+661F> <U+3082> <U+306A><U+3044> <U+7A7A> <U+306E> <U+4E2D><U+9014> <U+306B> <U+3001> <U+77ED><U+304F> <U+5782><U+308C> <U+3066> <U+3044><U+308B> <U+3070><U+304B><U+308A> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3059> <U+3002> <U+5FA1> <U+91C8><U+8FE6> <U+69D8> <U+306F> <U+6975><U+697D> <U+306E> <U+84EE><U+6C60> <U+306E> <U+3075><U+3061> <U+306B> <U+7ACB><U+3063> <U+3066> <U+3001> <U+3053><U+306E> <U+4E00><U+90E8><U+59CB><U+7D42> <U+3092> <U+3058><U+3063><U+3068> <U+898B> <U+3066> <U+3044><U+3089><U+3063><U+3057><U+3083><U+3044> <U+307E><U+3057> <U+305F> <U+304C> <U+3001> <U+3084><U+304C><U+3066> <U+9640><U+591A> <U+304C> <U+8840><U+306E><U+6C60> <U+306E> <U+5E95> <U+3078> <U+77F3> <U+306E> <U+3088><U+3046> <U+306B> <U+6C88><U+3093> <U+3067> <U+3057><U+307E><U+3044> <U+307E><U+3059> <U+3068> <U+3001> <U+60B2><U+3057> <U+305D><U+3046> <U+306A> <U+5FA1> <U+9854> <U+3092> <U+306A><U+3055> <U+308A> <U+306A><U+304C><U+3089> <U+3001> <U+307E><U+305F> <U+3076><U+3089><U+3076><U+3089> <U+5FA1> <U+6B69><U+304D> <U+306B> <U+306A><U+308A> <U+59CB><U+3081> <U+307E><U+3057> <U+305F> <U+3002> <U+81EA><U+5206> <U+3070><U+304B><U+308A> <U+5730><U+7344> <U+304B><U+3089> <U+306C><U+3051> <U+51FA><U+305D> <U+3046> <U+3068> <U+3059><U+308B> <U+3001> <U+9640><U+591A> <U+306E> <U+7121><U+6148><U+60B2> <U+306A> <U+5FC3> <U+304C> <U+3001> <U+305D><U+3046><U+3057><U+3066> <U+305D><U+306E> <U+5FC3> <U+76F8><U+5F53> <U+306A> <U+7F70> <U+3092> <U+3046><U+3051> <U+3066> <U+3001> <U+5143> <U+306E> <U+5730><U+7344> <U+3078> <U+843D><U+3061> <U+3066> <U+3057><U+307E><U+3063> <U+305F> <U+306E> <U+304C> <U+3001> <U+5FA1> <U+91C8><U+8FE6> <U+69D8> <U+306E> <U+5FA1> <U+76EE> <U+304B><U+3089> <U+898B><U+308B> <U+3068> <U+3001> <U+6D45><U+9593> <U+3057><U+304F> <U+601D><U+53EC> <U+3055> <U+308C> <U+305F> <U+306E> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3057><U+3087> <U+3046> <U+3002> <U+3057><U+304B><U+3057> <U+6975><U+697D> <U+306E> <U+84EE><U+6C60> <U+306E> <U+84EE> <U+306F> <U+3001> <U+5C11><U+3057> <U+3082> <U+305D><U+3093><U+306A> <U+4E8B> <U+306B> <U+306F> <U+9813><U+7740> <U+81F4><U+3057> <U+307E><U+305B> <U+3093> <U+3002> <U+305D><U+306E> <U+7389> <U+306E> <U+3088><U+3046> <U+306A> <U+767D><U+3044> <U+82B1> <U+306F> <U+3001> <U+5FA1> <U+91C8><U+8FE6> <U+69D8> <U+306E> <U+5FA1> <U+8DB3> <U+306E> <U+307E><U+308F><U+308A> <U+306B> <U+3001> <U+3086><U+3089><U+3086><U+3089> <U+843C> <U+3092> <U+52D5><U+304B><U+3057> <U+3066> <U+3001> <U+305D><U+306E> <U+307E><U+3093><U+4E2D> <U+306B> <U+3042><U+308B> <U+91D1><U+8272> <U+306E> <U+854A> <U+304B><U+3089> <U+306F> <U+3001> <U+4F55><U+3068><U+3082> <U+4E91><U+3048> <U+306A><U+3044> <U+597D><U+3044> <U+5302> <U+304C> <U+3001> <U+7D76><U+9593> <U+306A><U+304F> <U+3042><U+305F><U+308A> <U+3078> <U+6EA2><U+308C> <U+3066> <U+5C45><U+308A> <U+307E><U+3059> <U+3002> <U+6975><U+697D> <U+3082> <U+3082><U+3046> <U+5348> <U+306B> <U+8FD1><U+304F> <U+306A><U+3063> <U+305F> <U+306E> <U+3067> <U+3054><U+3056><U+3044> <U+307E><U+3057><U+3087> <U+3046> <U+3002>