Pack prettified data.frame of tokens

Pack a prettified data.frame of tokens into a new data.frame of corpus, which is compatible with the Text Interchange Formats.

pack(df, n = 1L, pull = "token", sep = "_", .collapse = " ")

Arguments

df
A prettified data.frame of tokens.

n
Integer internally passed to ngrams tokenizer function created of pipian::ngram_tokenizer()

pull
Column to be packed into text or ngrams body. Default value is `token`.

sep
Character scalar internally used as the concatenator of ngrams.

.collapse
This argument is passed to stringi::stri_join().

Value

A data.frame.

Text Interchange Formats (TIF)

The Text Interchange Formats (TIF) is a set of standards that allows R text analysis packages to target defined inputs and outputs for corpora, tokens, and document-term matrices.

Valid data.frame of tokens

The prettified data.frame of tokens here is a data.frame object compatible with the TIF.

A TIF valid data.frame of tokens are expected to have one unique key column (named `doc_id`) of each text and several feature columns of each tokens. The feature columns must contain at least `token` itself.

See also

Examples

xml <- ppn_parse_xml(system.file("sample.xml", package = "pipian"))
head(pack(xml))
#>   doc_id
#> 1      1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      text
#> 1 <U+3075><U+3068> <U+632F><U+308A><U+5411><U+304F> <U+3068> <U+3001> <U+305F><U+304F><U+3055><U+3093> <U+306E> <U+5473><U+65B9> <U+304C> <U+3044> <U+3066> <U+305F><U+304F><U+3055><U+3093> <U+306E> <U+512A><U+3057><U+3044> <U+4EBA><U+9593> <U+304C> <U+3044><U+308B> <U+3053><U+3068> <U+3092> <U+3001> <U+308F><U+3056><U+308F><U+3056> <U+81EA><U+5206> <U+306E> <U+8A95><U+751F> <U+65E5> <U+304C> <U+6765> <U+306A><U+3044> <U+3068> <U+6C17><U+4ED8><U+3051> <U+306A><U+3044> <U+81EA><U+5206> <U+3092> <U+596E><U+3044><U+7ACB><U+305F> <U+305B> <U+306A><U+304C><U+3089> <U+3082> <U+3001> <U+6BCE><U+65E5> <U+3053><U+3093><U+306A> <U+3001> <U+6E56> <U+306E> <U+3088><U+3046> <U+306A> <U+306A><U+3093> <U+306E> <U+5F15><U+3063><U+639B><U+304B><U+308A> <U+3082> <U+306A><U+3044> <U+3001> <U+843D><U+3061><U+3064><U+304D> <U+5012><U+3057> <U+3001> <U+97F3> <U+4E00><U+3064> <U+3082> <U+611F><U+3058> <U+3055><U+305B> <U+306A><U+3044> <U+4EBA><U+9593> <U+3067> <U+3044><U+308C><U+308B> <U+65B9> <U+306B> <U+61A7><U+308C> <U+3092> <U+6301><U+3066> <U+305F> <U+3068><U+3042><U+308B> 25 <U+6B73> <U+306E> <U+7729><U+3057><U+304D> <U+671D> <U+306E> <U+3053><U+3068> <U+3067><U+3057> <U+305F>